test: add tests for evaluate_polynomial in factor_theorem.py
