+++ Tests/test_Algebra.py
@@ -9,7 +9,6 @@ if root_dir not in sys.path:
 
 from Math.Algebra.Polynomials.polynomial import format_polynomial, evaluate_polynomial as evaluate_polynomial_poly
 from Math.Algebra.Linear_Equations.linear_eqn import linear_eqn
-from Math.Algebra.Polynomials.factor_theorem import evaluate_polynomial, check_factor
 
 def test_format_polynomial_basic():
     result = format_polynomial([2, 3, 4], [2, 1, 0])
@@ -68,78 +67,7 @@ def test_linear_eqn_vertical_line():
     # x1 == x2, expects ValueError
     with pytest.raises(ValueError, match="The x-coordinates cannot be the same \\(vertical line\\)."):
         linear_eqn(2, 3, 2, 7)
-class TestEvaluatePolynomial:
-    def test_evaluate_polynomial_basic(self):
-        """Test with a simple quadratic polynomial: x^2 + 2x + 1 at x=2"""
-        # (2)^2 + 2*(2) + 1 = 4 + 4 + 1 = 9
-        assert evaluate_polynomial([1, 2, 1], [2, 1, 0], 2) == 9
 
-    def test_evaluate_polynomial_zero_x(self):
-        """Test polynomial at x=0"""
-        # (0)^2 + 2*(0) + 1 = 1
-        assert evaluate_polynomial([1, 2, 1], [2, 1, 0], 0) == 1
-
-    def test_evaluate_polynomial_negative_x(self):
-        """Test polynomial at x=-1"""
-        # (-1)^2 + 2*(-1) + 1 = 1 - 2 + 1 = 0
-        assert evaluate_polynomial([1, 2, 1], [2, 1, 0], -1) == 0
-
-    def test_evaluate_polynomial_fractional_powers(self):
-        """Test with fractional powers (square root)"""
-        # 1 * (4)^0.5 = 2.0
-        assert evaluate_polynomial([1], [0.5], 4) == 2.0
-
-    def test_evaluate_polynomial_float_coefficients(self):
-        """Test with float coefficients and float x"""
-        # 1.5 * (2.0)^2 + 0.5 * (2.0)^0 = 1.5*4 + 0.5 = 6.0 + 0.5 = 6.5
-        assert evaluate_polynomial([1.5, 0.5], [2, 0], 2.0) == 6.5
-
-    def test_evaluate_polynomial_empty(self):
-        """Test with empty coefficients and powers"""
-        assert evaluate_polynomial([], [], 5) == 0
-
-    def test_evaluate_polynomial_negative_powers(self):
-        """Test with negative powers"""
-        # 4 * (2)^-1 + 2 * (2)^-2 = 4/2 + 2/4 = 2 + 0.5 = 2.5
-        assert evaluate_polynomial([4, 2], [-1, -2], 2) == 2.5
-def test_check_factor_true():
-    # P(x) = x^2 - 4x + 4, check if (x - 2) is a factor
-    # P(2) = 2^2 - 4*2 + 4 = 4 - 8 + 4 = 0 -> True
-    assert check_factor([1, -4, 4], [2, 1, 0], 2) is True
-
-def test_check_factor_false():
-    # P(x) = x^2 - 4x + 4, check if (x - 3) is a factor
-    # P(3) = 3^2 - 4*3 + 4 = 9 - 12 + 4 = 1 -> False
-    assert check_factor([1, -4, 4], [2, 1, 0], 3) is False
-
-def test_check_factor_linear():
-    # P(x) = 2x - 6, check if (x - 3) is a factor
-    # P(3) = 2*3 - 6 = 0 -> True
-    assert check_factor([2, -6], [1, 0], 3) is True
-
-def test_check_factor_linear_false():
-    # P(x) = 2x - 6, check if (x - 2) is a factor
-    # P(2) = 2*2 - 6 = -2 -> False
-    assert check_factor([2, -6], [1, 0], 2) is False
-
-def test_check_factor_cubic():
-    # P(x) = x^3 - 6x^2 + 11x - 6
-    # Factors are (x-1)(x-2)(x-3)
-    coefficients = [1, -6, 11, -6]
-    powers = [3, 2, 1, 0]
-    assert check_factor(coefficients, powers, 1) is True
-    assert check_factor(coefficients, powers, 2) is True
-    assert check_factor(coefficients, powers, 3) is True
-    assert check_factor(coefficients, powers, 4) is False
-
-def test_check_factor_float():
-    # P(x) = 2x^2 - x - 1
-    # Factors are (2x+1)(x-1) => x = -0.5, x = 1
-    coefficients = [2, -1, -1]
-    powers = [2, 1, 0]
-    assert check_factor(coefficients, powers, -0.5) is True
-    assert check_factor(coefficients, powers, 1.0) is True
-    assert check_factor(coefficients, powers, 0.5) is False
 def test_evaluate_polynomial_poly_basic():
     # P(x) = 2x^2 + 3x + 1
     # P(2) = 2(2^2) + 3(2) + 1 = 8 + 6 + 1 = 15
